requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
pandas>=1.1.0
scikit-image>=0.17.0
opencv-python>=4.5.0
albumentations>=1.0.0
yacs>=0.1.8
pytorch-msssim>=0.2.1
opencv-contrib-python>=4.5.0
requests>=2.25.0
gdown>=4.4.0
scikit-learn>=0.24.0
tqdm>=4.60.0 